Move some more of the plugins into folders
